Extract image loading from EditViewModel file
